+++ Legacy/bonej/src/test/java/org/bonej/plugins/SkeletonAnglesTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 
-import org.bonej.geometry.TestDataMaker;
+import org.bonej.util.TestDataMaker;
 import org.junit.Test;
 
 import ij.ImagePlus;
