+++ Legacy/bonej/src/test/java/org/bonej/plugins/StructureModelIndexTest.java
@@ -23,7 +23,7 @@ package org.bonej.plugins;
 
 import static org.junit.Assert.assertEquals;
 
-import org.bonej.geometry.TestDataMaker;
+import org.bonej.util.TestDataMaker;
 import org.junit.Ignore;
 import org.junit.Test;
